requirements.txt
streamlit
langchain
chromadb
unstructured
sentence-transformers
faiss-cpu
tiktoken
openai
pypdf
loguru
docx2txt
tiktoken